feat: Moved attack to PlayAttackAni for better handling.
