+++ Assets/UnZENity-Core/Scripts/Domain/Npc/Actions/AnimationActions/PlayAni.cs
@@ -1,6 +1,4 @@
-using GUZ.Core.Extensions;
 using GUZ.Core.Models.Container;
-using UnityEngine;
 
 namespace GUZ.Core.Domain.Npc.Actions.AnimationActions
 {
@@ -22,28 +20,6 @@ namespace GUZ.Core.Domain.Npc.Actions.AnimationActions
                 return;
             }
             ActionEndEventTime = PrefabProps.AnimationSystem.GetAnimationDuration(_animName);
-        }
-
-        // FIXME - Move to PlayAttackAni() to distinguish easily between normal animations and special attack ones.
-        public override void Tick()
-        {
-            base.Tick();
-            if (Action.Instance0 == null)
-                return;
-
-            var myPosition = NpcContainer.Go.transform.position;
-            var targetPosition = Action.Instance0.GetUserData()!.Go.transform.position;
-
-            // Consider only horizontal distance (ignore Y-axis)
-            var myPositionHorizontal = new Vector3(myPosition.x, 0, myPosition.z);
-            var targetPositionHorizontal = new Vector3(targetPosition.x, 0, targetPosition.z);
-            var distance = Vector3.Distance(myPositionHorizontal, targetPositionHorizontal);
-
-            if (distance <= 1f)
-            {
-                PrefabProps.AnimationSystem.StopAllAnimations();
-                IsFinishedFlag = true;
-            }
         }
     }
 }
