+++ Assets/UnZENity-Core/Scripts/Domain/Npc/Actions/AnimationActions/PlayAttackAni.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 13222f16466c4bcab2c24f7ecbe81da2
+timeCreated: 1764040097
